FIX: Screen shaking on mobile view navigation
Setting a border on the `fullscreen-table-wrapper` element causes a
small shake in mobile views, especially on topics with multiple tables.

This PR removes the border on the table entirely. This was an almost
invisible visual effect when selecting the table (small border width,
whole table is set to 50% opacity).

It also makes a small refactor, moving the styles of the button outside
of the hover state.