DEV: Introduce backwards-compat shims for older Discourse core (#57)
The changes in 8599bfcaff34993ab9e5e7d90edaabb1471a99b6 needed to be applied precisely in-step with the core changes from https://github.com/discourse/discourse/pull/23148. This is impossible to guarantee with a theme - `.discourse-compatibility` does not give us enough precision. So instead, we need some temporary backwards-compatibility shims so that the new relative-import paths still work on old discourse.